requirements.txt
pandas>=2.0.0
pandas-gbq>=0.19.0
google-cloud-bigquery>=3.14.0
google-api-core>=2.11.0
gspread>=5.12.0
gspread-dataframe>=3.3.0
zeep>=4.2.1
python-dateutil>=2.8.2
pyyaml>=6.0
google-auth>=2.23.0
python-dotenv>=1.0.0
